TRUNK-6463: Problem editing existing Complex Obs with the new Local File Storage Service TRUNK-6464: Test Context should not create new application data directory for each test

+++ api/src/test/java/org/openmrs/test/jupiter/BaseContextSensitiveTest.java
@@ -354,18 +354,20 @@ public abstract class BaseContextSensitiveTest {
 			runtimeProperties.setProperty(Environment.HBM2DDL_AUTO, "update");
 		}
 		
-		try {
-			File tempappdir = File.createTempFile("appdir-for-unit-tests-", "");
-			tempappdir.delete(); // so we can make it into a directory
-			tempappdir.mkdir(); // turn it into a directory
-			tempappdir.deleteOnExit(); // clean up when we're done with tests
-			
-			runtimeProperties.setProperty(OpenmrsConstants.APPLICATION_DATA_DIRECTORY_RUNTIME_PROPERTY, tempappdir
-			        .getAbsolutePath());
-			OpenmrsUtil.setApplicationDataDirectory(tempappdir.getAbsolutePath());
-		}
-		catch (IOException e) {
-			log.error("Unable to create temp dir", e);
+		String appDataDir = OpenmrsUtil.getApplicationDataDirectory();
+		if (appDataDir == null || appDataDir.contains("appdir-for-unit-tests-")) {
+			try {
+				File tempappdir = File.createTempFile("appdir-for-unit-tests-", "");
+				tempappdir.delete(); // so we can make it into a directory
+				tempappdir.mkdir(); // turn it into a directory
+				tempappdir.deleteOnExit(); // clean up when we're done with tests
+
+				runtimeProperties.setProperty(OpenmrsConstants.APPLICATION_DATA_DIRECTORY_RUNTIME_PROPERTY, tempappdir
+					.getAbsolutePath());
+				OpenmrsUtil.setApplicationDataDirectory(tempappdir.getAbsolutePath());
+			} catch (IOException e) {
+				log.error("Unable to create temp dir", e);
+			}
 		}
 		
 		return runtimeProperties;
